pyproject.toml
[project]
name = "shoppingagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-generativeai>=0.8.5",
    "langchain>=0.3.26",
    "litellm>=1.74.3",
    "openai-agents>=0.1.0",
    "py-dotenv>=0.1",
    "requests>=2.32.4",
    "streamlit>=1.46.1",
    "stripe>=12.3.0",
    "typing>=3.10.0.0",
    "typing-extensions>=4.14.1",
]
